Split useDispensing hook into three focused hooks following single responsibility principle
